fix(gateway): extract /auth add|readd|cancel dispatch out of gateway.ts (C1 ratchet)
The pushed branch left gateway.ts at 25097 lines; CI's merge-with-main
commit measured 25108, exceeding the ratchet hard limit (25053 + 50 grace
= 25103). Move the entire gateway-routed `/auth add | readd | cancel`
dispatch block (~80 lines, including PR C's readd precheck call site,
--replace threading, verb-noun success reply, and both-direction error
paths) out of gateway.ts into a new `handleAuthAddOrCancel` export in
auth-add-flow.ts — the module that already owns pendingAuthAddFlows and
the tmux/OAuth session lifecycle. Behavior is identical; the gateway now
passes an `isAdmin` boolean, chat/thread keys, and an html-reply callback.

gateway.ts: 25097 -> 25030 (merge-with-main projects to ~25041, well
under the 25103 ceiling, leaving headroom for the A->B->C combined merge).
